+++ examples/readme/src/main.rs
@@ -1,6 +1,9 @@
 use vyxen::prelude::*;
 
 fn main() {
+    #[cfg(target_arch = "wasm32")]
+    console_error_panic_hook::set_once();
+
     let mut game = Game::new();
     let mut scene = Scene::new();